real time mapping final version

+++ robot_demo/mapping.h
@@ -72,8 +72,8 @@ public:
     void extract_number(std::string& line);
     bool findPath(std::vector<POINT> &cesta,POINT start, POINT end);
 
-    std::vector< std::vector<uint8_t> > map;
-    std::vector< std::vector<uint8_t> > map_zaloha;
+    std::vector< std::vector<int> > map;
+    std::vector< std::vector<int> > map_zaloha;
     std::vector<POINT> new_maping_pose;
 
     pthread_mutex_t current_pose_lock;
